requirements.txt
streamlit==1.31.1
altair==4.2.2
pandas==2.2.2
gspread==6.1.2
google-auth==2.29.0
pdfminer.six==20231228
xlsxwriter==3.2.0
pymupdf==1.24.2